chore: refactor action task listing logic
- Add import statement for `convert` package in `action.go`
- Implement new `ListActionTasks` function in `action.go` to list repository action tasks with pagination and error handling
- Remove `actions.go` file and its contents including the `ListActionTasks` function
